refactor(demo/tooltip): Fix
